chore: reorganize folder structure for better project organization
