+++ src/classroom-app/src/pages/StudentDashboard.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import styles from "./StudentDashboard.module.css";
-import { get } from "./services/api";
-import Courses from "./Courses.jsx"
+import { get } from "../services/api.js";
+import Courses from "../components/Courses.jsx";
 
-function StudentDashboard () {
+function StudentDashboard() {
   const [selectedSection, setSelectedSection] = useState("dashboard");
   const [grades, setGrades] = useState([]);
   const [assignments, setAssignments] = useState([]);
@@ -75,17 +75,34 @@ function StudentDashboard () {
         <h1>Student Dashboard</h1>
         <nav className={styles.headerBtns}>
           <img src="\src\assets\dashboard.svg"></img>
-          <button onClick={() => setSelectedSection("dashboard")}>Dashboard</button>
+          <button onClick={() => setSelectedSection("dashboard")}>
+            Dashboard
+          </button>
           <img src="\src\assets\courses.svg"></img>
-          <button onClick={() => { setSelectedSection("courses"); fetchData("courses"); }}>
+          <button
+            onClick={() => {
+              setSelectedSection("courses");
+              fetchData("courses");
+            }}
+          >
             Courses
           </button>
           <img src="\src\assets\grades.svg"></img>
-          <button onClick={() => { setSelectedSection("grades"); fetchData("grades"); }}>
+          <button
+            onClick={() => {
+              setSelectedSection("grades");
+              fetchData("grades");
+            }}
+          >
             Grades
           </button>
           <img src="\src\assets\assign.svg"></img>
-          <button onClick={() => { setSelectedSection("assignments"); fetchData("assignments"); }}>
+          <button
+            onClick={() => {
+              setSelectedSection("assignments");
+              fetchData("assignments");
+            }}
+          >
             Assignments
           </button>
         </nav>
